+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All Notable changes to `laravel-state-machine` will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Added a MetadataStore to fetch metadata from graphs, states and transitions.
 
 ## [1.3.3] - 2019-04-18